requirements.txt
playwright
aiohttp
boto3
python-dotenv
